Extract sync implementation to service
